```text Refactor build process and add esbuild-plugin-fileloc
This commit refactors the build process in the `build.js` file. It introduces the `esbuild-plugin-fileloc` plugin to provide file location information during the build. The `buildBundle` function is renamed to `build` for clarity. The `WATCH` flag is replaced with the `--watch` command-line option. The client and server builds are now performed separately using the `esbuild.context` function. If the `--watch` option is provided, the builds are watched for changes. Otherwise, the builds are rebuilt once. The commit also updates the `package.json` file to include a `dev` script that runs the build in watch mode.